requirements.txt
streamlit==1.38.0 
langchain==0.3.3
python-dotenv==1.0.1
langchain-huggingface==0.1.0
langchain-google-genai==2.0.1
langchain-chroma==0.1.4
langchain-community==0.3.2
langchain-experimental==0.3.2
langgraph==0.2.38
langgraph-checkpoint-postgres==2.0.1
psycopg-pool==3.2.3
psycopg2==2.9.10